Move all auth lib content to core lib
