+++ libs/core/src/index.ts
@@ -1,11 +1,17 @@
 /*
  * Public API Surface of @flaps/core
  */
+export * from './lib/analytics';
 export * from './lib/api';
+export * from './lib/auth';
+export * from './lib/config';
 export * from './lib/file-upload';
 export * from './lib/models';
+export * from './lib/pipes';
 export * from './lib/testing';
+export * from './lib/ui';
+export * from './lib/utils';
 
 export * from './lib/load-script.service';
 export * from './lib/sidebar.service';
-export * from './lib/utils';
+export * from './lib/state.service';
