+++ libs/core/src/lib/api/google.service.ts
@@ -1,15 +1,14 @@
 import { Injectable } from '@angular/core';
 import { Observable, throwError } from 'rxjs';
 import { APIService } from './api.service';
-import { AuthTokens } from './models/user.model';
-import { BackendConfigurationService } from './backend-config.service';
+import { AuthTokens } from '../models';
+import { BackendConfigurationService } from '../config';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class GoogleService {
-
-  constructor(private api: APIService, private config: BackendConfigurationService) { }
+  constructor(private api: APIService, private config: BackendConfigurationService) {}
 
   getGoogleLoginUrl(): string {
     return this.config.getAPIURL() + `/auth/google/authorize`;
@@ -21,7 +20,7 @@ export class GoogleService {
       return throwError('Invalid state');
     }
     const data = { code: code };
-    return this.api.post(url, JSON.stringify(data), false );
+    return this.api.post(url, JSON.stringify(data), false);
   }
 
   private getLoginUrl(state: string): string | null {
@@ -29,8 +28,7 @@ export class GoogleService {
     try {
       // state is a base64 encoded json
       decoded = JSON.parse(atob(state));
-    }
-    catch {
+    } catch {
       return null;
     }
     const hasUrl = typeof decoded === 'object' && typeof decoded.login_url === 'string';
@@ -38,6 +36,6 @@ export class GoogleService {
   }
 
   private isSafeRedirect(redirectUrl: string) {
-    return redirectUrl.startsWith(this.config.getAPIURL())
+    return redirectUrl.startsWith(this.config.getAPIURL());
   }
 }
