+++ libs/core/src/lib/api/index.ts
@@ -1,2 +1,6 @@
+export * from './api.service';
+export * from './google.service';
+export * from './sdk.service';
+export * from './user.service';
 export * from './users.service';
 export * from './zone.service';
