+++ libs/core/src/lib/auth/token.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { of } from 'rxjs';
-import { APIService } from '../api.service';
-import { BackendConfigurationService } from '../backend-config.service';
+import { APIService } from '../api';
+import { BackendConfigurationService } from '../config';
 
 import { TokenService } from './token.service';
