+++ libs/core/src/lib/config/app.init.service.ts
