+++ libs/core/src/lib/config/index.ts
@@ -1,0 +1,3 @@
+export * from './app.init.service';
+export * from './backend-config.service';
+export * from './stf-config.module';
