+++ libs/core/src/lib/config/stf-config.module.ts
