+++ libs/core/src/lib/models/index.ts
@@ -1,2 +1,9 @@
+export * from './account.model';
+export * from './local.model';
+export * from './magic.model';
+export * from './permissions.model';
+export * from './region.model';
+export * from './user.model';
 export * from './users.model';
+export * from './welcome.model';
 export * from './zone.model';
