+++ libs/core/src/lib/models/welcome.model.ts
