+++ libs/core/src/lib/pipes/asset-url.pipe.ts
@@ -1,15 +1,13 @@
-import { Pipe, PipeTransform, Inject } from '@angular/core';
-import { BackendConfigurationService } from '../backend-config.service';
+import { Pipe, PipeTransform } from '@angular/core';
+import { BackendConfigurationService } from '../config';
 
 @Pipe({
-  name: 'assetUrl'
+  name: 'assetUrl',
 })
 export class AssetUrlPipe implements PipeTransform {
-
   constructor(private config: BackendConfigurationService) {}
 
   transform(value: string): string {
     return (this.config.getBaseAssetUrl() || '') + value;
   }
-
 }
