+++ libs/core/src/lib/pipes/filter.pipe.ts
