+++ libs/core/src/lib/pipes/floor.pipe.ts
