+++ libs/core/src/lib/pipes/index.ts
@@ -1,12 +1,13 @@
+export * from './asset-url.pipe';
 export * from './bold.pipe';
-export * from './size.pipe';
-export * from './safe.pipe';
+export * from './floor.pipe';
+export * from './filter.pipe';
 export * from './get-by-id.pipe';
 export * from './get-file.pipe';
 export * from './keys.pipe';
+export * from './safe.pipe';
+export * from './size.pipe';
 export * from './time-ago.pipe';
-export * from './floor.pipe';
-export * from './filter.pipe';
 export * from './trim.pipe';
 export * from './sanitize-tags.pipe';
-export * from './asset-url.pipe';
+export * from './stf-pipes.module';
