+++ libs/core/src/lib/pipes/safe.pipe.ts
