+++ libs/core/src/lib/pipes/trim.pipe.ts
