+++ libs/core/src/lib/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './splash-screen.service';
