+++ libs/core/src/lib/ui/splash-screen.service.ts
